Massive refactoring and changes
Added pystatsd to the cookbook, also I refactored all of the python
related code to work with the opscode python cookbook. The uwsgi stuff
wasn't working for me either so I removed the providers and just created
a basic template setup. Ideally this would be changed to use a uwsgi
cookbook of some sort. There's a good bit of abstraction that could be
done and it still needs run through foodcritic at some point.